Speculative password import prompt crash fix (#3542)
Task/Issue URL:
https://app.asana.com/0/1203822806345703/1208648341254407/f
CC: @dbajpeyi 

**Description**:

There's a crash in moderate numbers which is occurring during the
password import prompt flow. This is speculative fix for that.

The reason seemed to be that the replyHandler was being called too many
times. However:

- The docs for
`userContentController:didReceiveScriptMessage:replyHandler:` say "The
replyHandler can be called at most once. If the replyHandler is
deallocated before it is called, the Promise will be rejected with a
JavaScript Error object
with an appropriate message indicating the handler was never called."
- In the case of startCredentialsImportFlow where the crash is
occurring, the JS is not awaiting. So there is no need to call the
replyHandler.

**Steps to test this PR:**
Just smoke tests for the feature:
https://app.asana.com/0/1142021229838617/1208543152554655/f

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)