PIR E2E tests GitHub actions (#3528)
Task/Issue URL:
https://app.asana.com/0/72649045549333/1208698939169508/f
Tech Design URL:
CC:

**Description**:
Adds the yml file to run the new PIR e2e tests. It should trigger on:
- PRs against release and hotfix branches
- Nightly
- If the new PR checkbox is ticked (See below)

Also slips in a minor change to the already merged tests to prevent a
potential crash which offered less useful error messaging than if the
expectation wasn't fulfilled

**Optional E2E tests**:
- [x] Run PIR E2E tests
Check this to run the Personal Information Removal end to end tests. If
updating CCF, or any PIR related code, tick this.

**Steps to test this PR**:
1. Check the yml that everything looks sensible
2. Check that the tests pass against this PR [Note: The staging backend
is currently down, which the tests are reliant on, so at the moment they
fail]
3. I've tested with a variety of setups with branches to test that
trigger, but if keen you could also try making a release of hotfix
branch to test against.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)