macOS 13/14 UI test compilation fix (#3569)
Task/Issue URL: https://app.asana.com/0/1201037661562251/1208772745284327/f
Tech Design URL:
CC:

Description:

This PR fixes UI test compilation on macOS 13/14.

The issue was that the UI test suite was linking BSK, which causes Xcode to then compile it - but, we can't allow it to do that since GRDB was built with a newer version of Xcode.