CRCID Support when reporting crashes (#3653)
Task/Issue URL:
https://app.asana.com/0/1208592102886666/1208759541597499/f
Tech Design URL:
https://app.asana.com/0/1208592102886666/1208660326715650/f

**Description**:
DO NOT MERGE - this is a draft for input, not ready to go live yet.

MacOS support for crash report cohort IDs (CRCIDs) when sending crash
reports.

**Optional E2E tests**:
- [ ] Run PIR E2E tests
Check this to run the Personal Information Removal end to end tests. If
updating CCF, or any PIR related code, tick this.

**Steps to test this PR**:
To cause and report a crash:

1. Launch the app and force a crash, which can be done from Debug (menu)
→ Simulate Crash → fatalError (or similar).
2. Launch the app again (easiest with a debugger)
3. You will be prompted to submit the crash. When submitting, watch logs
for “crcid” and “crash” and you should see output from these changes
confirming a CRCID is received and updated.
4. Repeat steps 2-3 again, and you should see that the CRCID received
initially is what’s submitted with the next crash.