Add Freemium PIR banner to HTML New Tab Page (#3706)
Task/Issue URL: https://app.asana.com/0/72649045549333/1208973375734052/f

Description:
This change adds support for displaying Freemium PIR banner in the HTML New Tab Page.
The logic of FreemiumDBPPromotionViewCoordinator is updated so that the viewModel is
a published property, updated whenever isHomePagePromotionVisible changes (which also
means that it's nullified whenever the user dismisses or actions a banner).