Reload webview when tab crashes (#3766)
Task/Issue URL:
https://app.asana.com/0/1201048563534612/1209189376094874/f
Tech Design URL:
CC:

**Description**
Silently reload the tab when we get a crash and no form is present. The
plan is to get this only for internal users and test for a few weeks.

**Steps to test this PR**:
It is hard to test, but opening Sharepoint in a tab and leaving it in
the background for some time could cause a termination. You could attach
the debugger and check if it was terminated and the code was executed.

**Definition of Done**:

* [x] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)