Sync: Fire accountRemoved KV store reason when no account (#3782)
Task/Issue URL:
https://app.asana.com/0/1202926619870900/1209014391063056/f

**Description**:

As part of [✓ Send pixels for account removal + decoding
issues](https://app.asana.com/0/1199230911884351/1208723035104886/f),
pixels were added indicating what the reason for the Sync account being
removed from the Keychain was.

However, the pixel
`sync_account_removed_reason_not-set-on-key-value-store` is being sent
even when there was no account stored in the keychain in the first
place. This could be obscuring a problem.

To make this Pixel more useful, we should update it so that it's only
sent when there was an account stored in the keychain.

**Steps to test this PR**:
Hard to test the pixel itself as reporting on a broken state, but just
do a quick smoke test of Sync on both platforms.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Definition of Done**:

* [x] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)